+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 10.6.0
-- Added setting to disable the shipping callback for express checkouts.
-- Fixes an issue, where the shiiping callback required the store-api to be exposed.
+- Added setting to disable the shipping callback for express checkouts
+- Fixes an issue, where the shiiping callback required the store-api to be exposed
+- Fixes an issue, where the cart was not correctly detected as free, resulting in errors during the PayPal checkout (shopware/SwagPayPal#591)
 
 # 10.5.0
 - Added Austria to the countries where Pay Later is available